% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{build_xpath}
\alias{build_xpath}
\title{Build and xpath string, possibly using the user specified namespace
declaration.}
\usage{
build_xpath(x, prefix = ".//", ns = "")
}
\arguments{
\item{x}{character one or more path segments}

\item{prefix}{character either "./", ".//" or "//" or if NA or NA then ignored}

\item{ns}{if "" or NA then ignored otherwise it is appended to each segment}
}
\value{
xpath descriptor
}
\description{
Build and xpath string, possibly using the user specified namespace
declaration.
}
