% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Catalog.R
\docType{class}
\name{CatalogRefClass-class}
\alias{CatalogRefClass-class}
\alias{CatalogRefClass}
\title{An catalogRef representation that subclasses from ThreddsNodeRefClass}
\description{
An catalogRef representation that subclasses from ThreddsNodeRefClass
}
\section{Fields}{

\describe{
\item{\code{name}}{character}

\item{\code{href}}{character relative link}

\item{\code{title}}{character}

\item{\code{type}}{character}

\item{\code{ID}}{character}
}}

\section{Methods}{

\describe{
\item{\code{initialize(x, verbose = FALSE, n_tries = 3)}}{x may be url or xml2::xml_node}

\item{\code{show(prefix = "")}}{show the content of the class}
}}

\concept{Catalog}
