% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Catalog.R
\name{CatalogRefClass_get_catalog}
\alias{CatalogRefClass_get_catalog}
\title{Retrieve the top catalog this catalog points to}
\value{
TopCatalogRefClass or NULL
}
\description{
Retrieve the top catalog this catalog points to
}

