% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\docType{class}
\name{DatasetRefClass-class}
\alias{DatasetRefClass-class}
\alias{DatasetRefClass}
\title{A direct Dataset representation that subclasses from ThreddsNodeRefClass}
\description{
A direct Dataset representation that subclasses from ThreddsNodeRefClass
}
\section{Fields}{

\describe{
\item{\code{dataSize}}{numeric size in bytes}

\item{\code{date}}{character}

\item{\code{urlPath}}{character relative URL, use url or get_url() instead}
}}

\section{Methods}{

\describe{
\item{\code{show(prefix = "")}}{show the contents}
}}

\seealso{
Other Thredds: \code{\link{DatasetsRefClass-class}},
  \code{\link{ServiceRefClass-class}},
  \code{\link{ThreddsNodeRefClass_GET}},
  \code{\link{ThreddsNodeRefClass_get_url}},
  \code{\link{ThreddsNodeRefClass_unames}}
}
\concept{Thredds}
