% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{DatasetRefClass_get_url}
\alias{DatasetRefClass_get_url}
\title{Retrieve the URL for a dataset}
\arguments{
\item{service}{character or NULL.  If not NULL then substitute the value
of \code{replace} with this.  OPeNDAP is "thredds/dodsC".}
}
\value{
character
}
\description{
Datasets may not be aware of the the way they are served (OPeNDAP, WMS, NCML, etc)
Typically this maybe known at the TopCatalog that reference the dataset.
This function provides a convenient way to form the URL for data access rather than
simple HMTL viewing.
}

