% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\name{DatasetsRefClass_get_collection}
\alias{DatasetsRefClass_get_collection}
\title{Retrieve the datasets from a dataset collection}
\value{
a list of zero or more DatasetRefClass
}
\description{
Retrieve the datasets from a dataset collection
}

