% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_catalog}
\alias{get_catalog}
\title{Retrieve a catalog}
\usage{
get_catalog(uri, ...)
}
\arguments{
\item{uri}{the URI of the catalog}

\item{...}{further arguments for parse_node}
}
\value{
ThreddsNodeRefClass or subclass or NULL
}
\description{
Retrieve a catalog
}
