% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_xml_ns}
\alias{get_xml_ns}
\title{Retrieve the namespaces for a resource}
\usage{
get_xml_ns(uri)
}
\arguments{
\item{uri}{the URI of the catalog}
}
\value{
the output of \code{\link[xml2]{xml_ns}}
}
\description{
Retrieve the namespaces for a resource
}
