% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{grepl_it}
\alias{grepl_it}
\title{Determine if a vector of names match the greplargs}
\usage{
grepl_it(x, greplargs = NULL)
}
\arguments{
\item{x}{a vector of names}

\item{greplargs}{NULL, vector or list}
}
\value{
logical vector
}
\description{
Determine if a vector of names match the greplargs
}

