% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{parse_node}
\alias{parse_node}
\title{Convert a node to an object inheriting from ThreddsNode}
\usage{
parse_node(node, url = NULL, verbose = FALSE, encoding = "UTF-8", ...)
}
\arguments{
\item{node}{xml2::xml_node or an httr::response object}

\item{url}{character, optional url if a catalog or direct dataset}

\item{verbose}{logical, by default FALSE}

\item{encoding}{character, by default UTF-8}

\item{...}{further arguments for instantiation of classes (such as ns = "foo")}
}
\value{
ThreddsNode class object or subclass
}
\description{
Convert a node to an object inheriting from ThreddsNode
}
