% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Service.R
\docType{class}
\name{ServiceRefClass-class}
\alias{ServiceRefClass-class}
\alias{ServiceRefClass}
\title{An Service representation that subclasses from ThreddsNodeRefClass}
\description{
An Service representation that subclasses from ThreddsNodeRefClass
}
\section{Fields}{

\describe{
\item{\code{name}}{character}

\item{\code{serviceType}}{character}

\item{\code{base}}{character base url}
}}

\section{Methods}{

\describe{
\item{\code{initialize(x, verbose = FALSE, n_tries = 3, ns = "")}}{x may be url or xml2::xml_node}

\item{\code{show(prefix = "")}}{show the content of the class}
}}

\seealso{
Other Thredds: \code{\link{DatasetRefClass-class}},
  \code{\link{DatasetsRefClass-class}},
  \code{\link{ThreddsNodeRefClass_GET}},
  \code{\link{ThreddsNodeRefClass_browse}},
  \code{\link{ThreddsNodeRefClass_get_url}},
  \code{\link{ThreddsNodeRefClass_unames}},
  \code{\link{ThreddsNodeRefClass_xpath}}
}
\concept{Thredds}
