% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tds_list_datasets.R
\name{tds_list_datasets}
\alias{tds_list_datasets}
\title{Get a list of available datasets on a THREDDS data server.}
\usage{
tds_list_datasets(thredds_url, recursive = FALSE)
}
\arguments{
\item{thredds_url}{A string providing the URL of a THREDDS server,
usually ending with '/thredds/'.}

\item{recursive}{Should the function recurse into nested THREDDS catalogs.
Defaults to 'FALSE'.}
}
\value{
A data_frame containing dataset names and paths.
}
\description{
Get a list of available datasets on a THREDDS data server.
}
\examples{
library(thredds)
tds_list_datasets(thredds_url = "https://cida.usgs.gov/thredds/")
tds_list_datasets(thredds_url = "http://thredds.northwestknowledge.net:8080/thredds/")
}
