% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tds_list_services.R
\name{tds_list_services}
\alias{tds_list_services}
\title{List THREDDS dataset service endpoints}
\usage{
tds_list_services(dataset_url)
}
\arguments{
\item{dataset_url}{The url of a THREDDS dataset,
perhaps retrieved using tds_list_datasets.}
}
\value{
A data_frame containing dataset service endpoints and paths.
}
\description{
List THREDDS dataset service endpoints
}
\examples{
library(thredds)
thredds_data <- "https://cida.usgs.gov/thredds/catalog.html?dataset=cida.usgs.gov/prism_v2"
tds_list_services(dataset_url = thredds_data)
}
