% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tds_ncss_list_vars.R
\name{tds_ncss_list_vars}
\alias{tds_ncss_list_vars}
\title{List available variables of a THREDDS dataset via a NetcdfSubset service}
\usage{
tds_ncss_list_vars(ncss_url)
}
\arguments{
\item{ncss_url}{The url of a THREDDS NetcdfSubset service,
perhaps retrieved using tds_list_services.}
}
\value{
A data frame of variables and their attributes
}
\description{
List available variables of a THREDDS dataset via a NetcdfSubset service
}
\examples{
library(thredds)
ncss <- "https://cida.usgs.gov/thredds/ncss/macav2metdata_monthly_historical/dataset.html"
tds_ncss_list_vars(ncss_url = ncss)

}
