% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threddr_list_datasets.R
\name{threddr_list_datasets}
\alias{threddr_list_datasets}
\title{Get a list of available datasets on a THREDDS data server.}
\usage{
threddr_list_datasets(thredds_url, recursive = FALSE, datasets_only = FALSE)
}
\arguments{
\item{thredds_url}{A string providing the URL of a THREDDS server,
usually ending with '/thredds/'.}

\item{recursive}{Should the function recurse into nested THREDDS catalogs.
Defaults to 'FALSE'.}

\item{datasets_only}{Should the function only return datasets
(and exclude, for example, nested THREDDS catalogs).
Defaults to 'FALSE'.}
}
\value{
A data_frame containing dataset names and paths.
}
\description{
Get a list of available datasets on a THREDDS data server.
}
\examples{
threddr_list_datasets(thredds_url = "https://cida.usgs.gov/thredds/")
threddr_list_datasets(thredds_url = "http://thredds.northwestknowledge.net:8080/thredds/")
}
