% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threddr_list_services.R
\name{threddr_list_services}
\alias{threddr_list_services}
\title{List THREDDS dataset service endpoints}
\usage{
threddr_list_services(dataset_url)
}
\arguments{
\item{dataset_url}{The url of a THREDDS dataset,
perhaps retrieved using threddr_list_datasets.}
}
\value{
A data_frame containing dataset service endpoints and paths.
}
\description{
List THREDDS dataset service endpoints
}
\examples{
threddr_list_services(dataset_url = "https://cida.usgs.gov/thredds/catalog.html?dataset=cida.usgs.gov/prism_v2")
}
