% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threddr_ncss_download.R
\name{threddr_ncss_download}
\alias{threddr_ncss_download}
\title{Download a THREDDS dataset via a NetcdfSubset service}
\usage{
threddr_ncss_download(ncss_url, out_dir = "./", bbox = NULL, vars = NULL,
  add_args = NULL, overwrite = TRUE, ...)
}
\arguments{
\item{ncss_url}{The url of a THREDDS NetcdfSubset service,
perhaps retrieved using threddr_list_services.}

\item{out_dir}{A directory in which to download the dataset.
Defaults to the current working directory.}

\item{bbox}{An optional object of class 'bbox' from which to construct a bounding box.
Must be in EPSG:4326.
Defaults to NULL, which will download the entire NetCDF dataset.}

\item{vars}{A character vector of variable names,
perhaps retrieved using threddr_dodsC_list_vars.
Defaults to all available variables.}

\item{add_args}{A named list of additional NetcdfSubset arguments for the URL.}

\item{overwrite}{A logical; whether to overwrite the downloaded file if it exists.
Defaults to 'TRUE'.}

\item{...}{Additional arguments passed on to httr::GET.}
}
\value{
The path to the downloaded dataset.
}
\description{
Download a THREDDS dataset via a NetcdfSubset service
}
\examples{
\dontrun{
  threddr_ncss_download(ncss_url = "https://cida.usgs.gov/thredds/ncss/macav2metdata_monthly_historical/dataset.html",
                        bbox = st_bbox(c(xmin = -116, xmax = -115, ymax = 44, ymin = 45)),
                        vars = c("huss_BNU-ESM_r1i1p1_historical","huss_CCSM4_r6i1p1_historical"))
}
}
