% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thredds.R
\docType{class}
\name{ThreddsNodeRefClass-class}
\alias{ThreddsNodeRefClass-class}
\alias{ThreddsNodeRefClass}
\title{An base representation that other nodes subclass from}
\description{
An base representation that other nodes subclass from
}
\section{Fields}{

\describe{
\item{\code{url}}{character - possibly wrong but usually right!}

\item{\code{node}}{xml2::xml_node}

\item{\code{handle}}{httr::handle object}

\item{\code{verbose_mode}}{logical}

\item{\code{tries}}{numeric}
}}

\section{Methods}{

\describe{
\item{\code{initialize(x, verbose = FALSE, n_tries = 3)}}{x may be url or xml2::xml_node}
}}

\concept{Thredds2}
