% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thredds.R
\name{ThreddsNodeRefClass_browse}
\alias{ThreddsNodeRefClass_browse}
\title{Browse the URL if possible}
\value{
NULL invisibly
}
\description{
Browse the URL if possible
}
\seealso{
Other Thredds: \code{\link{DatasetRefClass-class}},
  \code{\link{DatasetsRefClass-class}},
  \code{\link{ServiceRefClass-class}},
  \code{\link{ThreddsNodeRefClass_GET}},
  \code{\link{ThreddsNodeRefClass_get_url}},
  \code{\link{ThreddsNodeRefClass_unames}},
  \code{\link{ThreddsNodeRefClass_xpath}}
}
\concept{Thredds}
