% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thredds.R
\name{ThreddsNodeRefClass_GET}
\alias{ThreddsNodeRefClass_GET}
\title{Retrieve a node of the contents at this nodes URL}
\value{
ThreddsNodeRefClass or subclass or NULL
}
\description{
Retrieve a node of the contents at this nodes URL
}
\seealso{
Other Thredds: \code{\link{DatasetRefClass-class}},
  \code{\link{DatasetsRefClass-class}},
  \code{\link{ServiceRefClass-class}},
  \code{\link{ThreddsNodeRefClass_browse}},
  \code{\link{ThreddsNodeRefClass_get_url}},
  \code{\link{ThreddsNodeRefClass_unames}},
  \code{\link{ThreddsNodeRefClass_xpath}}
}
\concept{Thredds}
