% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thredds.R
\name{ThreddsNodeRefClass_get_url}
\alias{ThreddsNodeRefClass_get_url}
\title{Retrieve the url of this node (mostly gets an override by subclasses?)}
\value{
character url (possibly invalid)
}
\description{
Retrieve the url of this node (mostly gets an override by subclasses?)
}
\seealso{
Other Thredds: \code{\link{DatasetRefClass-class}},
  \code{\link{DatasetsRefClass-class}},
  \code{\link{ServiceRefClass-class}},
  \code{\link{ThreddsNodeRefClass_GET}},
  \code{\link{ThreddsNodeRefClass_browse}},
  \code{\link{ThreddsNodeRefClass_unames}},
  \code{\link{ThreddsNodeRefClass_xpath}}
}
\concept{Thredds}
