% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thredds.R
\name{ThreddsNodeRefClass_xpath}
\alias{ThreddsNodeRefClass_xpath}
\title{Build an xpath specification}
\arguments{
\item{x}{character one or more path segments}

\item{ns}{if "" or NA then ignored otherwise it is appended to each segment}

\item{...}{further arguments for \code{\link{build_xpath}}}
}
\value{
xpath descriptor
}
\description{
Build an xpath specification
}
\seealso{
Other Thredds: \code{\link{DatasetRefClass-class}},
  \code{\link{DatasetsRefClass-class}},
  \code{\link{ServiceRefClass-class}},
  \code{\link{ThreddsNodeRefClass_GET}},
  \code{\link{ThreddsNodeRefClass_browse}},
  \code{\link{ThreddsNodeRefClass_get_url}},
  \code{\link{ThreddsNodeRefClass_unames}}
}
\concept{Thredds}
