% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopCatalog.R
\docType{class}
\name{TopCatalogRef-class}
\alias{TopCatalogRef-class}
\alias{TopCatalogRefClass}
\title{An catalog representation that sublcasses from ThreddsNodeRefClass}
\description{
An catalog representation that sublcasses from ThreddsNodeRefClass
}
\section{Methods}{

\describe{
\item{\code{show(prefix = "")}}{show the content of the class}
}}

\concept{Thredds TopCatalog}
