% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopCatalog.R
\name{TopCatalogRefClass_get_catalogs}
\alias{TopCatalogRefClass_get_catalogs}
\title{Get a list of CatalogRef Children}
\value{
a list of CatalogRefClass, possibly NULL
}
\description{
Get a list of CatalogRef Children
}
\seealso{
Other TopCatalog: \code{\link{TopCatalogRefClass_get_datasets}}
}
