% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopCatalog.R
\name{TopCatalogRefClass_get_datasets}
\alias{TopCatalogRefClass_get_datasets}
\title{Get a list of Dataset Children}
\value{
a list of DatasetRefClass, possibly NULL
}
\description{
Get a list of Dataset Children
}
\seealso{
Other TopCatalog: \code{\link{TopCatalogRefClass_get_catalogs}}
}
