% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopCatalog.R
\name{TopCatalogRefClass_list_services}
\alias{TopCatalogRefClass_list_services}
\title{List the services (if any)}
\arguments{
\item{xpath}{character xpath representation}
}
\value{
list of zero or more character vectors
}
\description{
List the services (if any)
}
\seealso{
Other TopCatalog: \code{\link{TopCatalogRefClass_get_catalog_names}},
  \code{\link{TopCatalogRefClass_get_catalogs}},
  \code{\link{TopCatalogRefClass_get_dataset_names}},
  \code{\link{TopCatalogRefClass_get_datasets}},
  \code{\link{TopCatalogRefClass_parse_dataset}}
}
\concept{TopCatalog}
