% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopCatalog.R
\name{TopCatalogRefClass_parse_dataset}
\alias{TopCatalogRefClass_parse_dataset}
\title{Parse a catlog node}
\arguments{
\item{x}{xml_node}

\item{x}{xml_node}
}
\value{
CatalogRefNode

DatasetRefNode
}
\description{
Parse a catlog node

Parse a dataset node
}
\seealso{
Other TopCatalog: \code{\link{TopCatalogRefClass_get_catalog_names}},
  \code{\link{TopCatalogRefClass_get_catalogs}},
  \code{\link{TopCatalogRefClass_get_dataset_names}},
  \code{\link{TopCatalogRefClass_get_datasets}}

Other TopCatalog: \code{\link{TopCatalogRefClass_get_catalog_names}},
  \code{\link{TopCatalogRefClass_get_catalogs}},
  \code{\link{TopCatalogRefClass_get_dataset_names}},
  \code{\link{TopCatalogRefClass_get_datasets}}
}

