% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{xml_children_names}
\alias{xml_children_names}
\title{Get the names of children}
\usage{
xml_children_names(x, unique_only = TRUE)
}
\arguments{
\item{x}{xml2::xml_node}

\item{unique_only}{logical if TRUE remove duplicates}
}
\value{
zero or more child names.  If none an empty character string is returned
}
\description{
Get the names of children
}
