% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{xml_id}
\alias{xml_id}
\title{Retrieve an ID value for a node from it's attributes.}
\usage{
xml_id(x, atts = c("name", "title", "ID", "urlPath", "href"))
}
\arguments{
\item{x}{xml node or a named character vector as per \code{xml_attrs}}

\item{atts}{character, ordered vector of attribute names to use as an ID value
As the list is stepped through if an attribute is missing or empty character
then advance to the next, otherwise return that value}
}
\value{
character identifier, possibly an empty character (\code{character()})
}
\description{
Retrieve an ID value for a node from it's attributes.
}
