% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{build_xpath}
\alias{build_xpath}
\title{Build and xpath string, possibly using the user specified namespace
prefix.}
\usage{
build_xpath(x, prefix = "d1", select = ".//")
}
\arguments{
\item{x}{character one or more path segments}

\item{prefix}{character by default "d1" prepended to each of the segements
in \code{x}.  If NA or length is 0 then ignore.}

\item{select}{charcater, by default search anywhere in the current node with ".//"}
}
\value{
xpath descriptor
}
\description{
Build and xpath string, possibly using the user specified namespace
prefix.
}
