% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ServiceNode.R
\name{ServiceNode}
\alias{ServiceNode}
\title{A simple class for parsing and holdoing service info}
\description{
A Service representation that subclasses from ThreddsNode
}
\note{
For examples see \link{CatalogNode}
}
\section{Super class}{
\code{\link[thredds:ThreddsNode]{thredds::ThreddsNode}} -> \code{ServiceNode}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character}

\item{\code{serviceType}}{character}

\item{\code{base}}{character base url}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ServiceNode-new}{\code{ServiceNode$new()}}
\item \href{#method-ServiceNode-print}{\code{ServiceNode$print()}}
\item \href{#method-ServiceNode-clone}{\code{ServiceNode$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="thredds" data-topic="ThreddsNode" data-id="GET"><a href='../../thredds/html/ThreddsNode.html#method-ThreddsNode-GET'><code>thredds::ThreddsNode$GET()</code></a></span></li>
<li><span class="pkg-link" data-pkg="thredds" data-topic="ThreddsNode" data-id="browse"><a href='../../thredds/html/ThreddsNode.html#method-ThreddsNode-browse'><code>thredds::ThreddsNode$browse()</code></a></span></li>
<li><span class="pkg-link" data-pkg="thredds" data-topic="ThreddsNode" data-id="children_names"><a href='../../thredds/html/ThreddsNode.html#method-ThreddsNode-children_names'><code>thredds::ThreddsNode$children_names()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceNode-new"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceNode-new}{}}}
\subsection{Method \code{new()}}{
initialize an instance of ServiceNode
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceNode$new(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{url or xml2::xml_node}

\item{\code{...}}{arguments for superclass initialization}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceNode-print"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceNode-print}{}}}
\subsection{Method \code{print()}}{
print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceNode$print(prefix = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prefix}}{character, to be printed before each line of output (like spaces)}

\item{\code{...}}{other arguments for superclass}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceNode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceNode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
