% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thredds2-package.R
\docType{package}
\name{thredds}
\alias{thredds}
\alias{thredds-package}
\title{thredds2: A crawler for navigating THREDDS catalogs}
\description{
A limited crawler for programmatically navigating THREDDS catalogs.
}

