% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thredds-package.R
\docType{package}
\name{thredds}
\alias{thredds}
\alias{thredds-package}
\title{thredds: Crawler for Navigating THREDDS Catalogs}
\description{
A limited crawler for programmatically navigating THREDDS catalogs.
}
