% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopCatalog.R
\name{TopCatalogRefClass_get_catalog_names}
\alias{TopCatalogRefClass_get_catalog_names}
\title{Get a list of CalatogRef child names}
\value{
character vector
}
\description{
Not all catalog refs have a name attribute.  When absent we try to substitute
 title, ID and finally href in that order.
}
\seealso{
Other TopCatalog: \code{\link{TopCatalogRefClass_get_catalogs}},
  \code{\link{TopCatalogRefClass_get_dataset_names}},
  \code{\link{TopCatalogRefClass_get_datasets}},
  \code{\link{TopCatalogRefClass_list_services}},
  \code{\link{TopCatalogRefClass_parse_dataset}}
}

